refactor: add LottoIOHandler to seperate IO logic
