+++ src/components/LottoController.js
@@ -1,19 +1,19 @@
-import { Console, Random } from '@woowacourse/mission-utils';
-import {
-  DELIMETER,
-  InputPrompts,
-  Lotto,
-  OutputMessages,
-  Prize,
-  PrizeMoney,
-} from '../resources/Constants.js';
+import { Random } from '@woowacourse/mission-utils';
+import { Lotto } from '../resources/Constants.js';
 import purchaseAmountValidator from '../validation/purchaseAmountValidator.js';
 import { bonusNumberValidator } from '../validation/bonusNumberValidator.js';
+import LottoIOHandler from './LottoIOHandler.js';
+import isEmpty from '../utils/isEmpty.js';
 
 class LottoController {
   #purchaseAmount;
   #lottoTickets;
   #bonusNumber;
+  ioHandler;
+
+  constructor() {
+    this.ioHandler = new LottoIOHandler();
+  }
 
   getPurchaseAmount() {
     return this.#purchaseAmount;
@@ -37,20 +37,22 @@ class LottoController {
     this.#bonusNumber = Number(bonusNumber);
   }
 
-  isEmptyPurchaseAmount() {
-    return this.#purchaseAmount === undefined || this.#purchaseAmount === null;
-  }
-
-  isEmptyLottoTickets() {
-    return this.#lottoTickets === undefined || this.#lottoTickets === null;
-  }
-
   #sortAscending(lottoTickets) {
     return lottoTickets.map((ticket) => ticket.slice().sort((a, b) => a - b));
   }
 
+  async promptPurchaseAmount() {
+    const purchaseAmount = await this.ioHandler.promptPurchaseAmount();
+    this.setPurchaseAmount(purchaseAmount);
+  }
+
+  async promptBonusNumber() {
+    const bonusNumber = await this.ioHandler.promptBonusNumber();
+    this.setBonusNumber(bonusNumber);
+  }
+
   generateLottoTickets() {
-    if (!this.isEmptyPurchaseAmount()) {
+    if (!isEmpty(this.#purchaseAmount)) {
       const ticketCount = this.#purchaseAmount / 1000;
 
       const lottoTickets = Array.from({ length: ticketCount }, () =>
@@ -62,42 +64,6 @@ class LottoController {
       );
 
       this.#lottoTickets = this.#sortAscending(lottoTickets);
-    }
-  }
-
-  displayLottoTickets() {
-    if (!this.isEmptyLottoTickets()) {
-      const ticketCount = this.#purchaseAmount / 1000;
-
-      Console.print(OutputMessages.PURCHASE_MESSAGE(ticketCount));
-
-      this.#lottoTickets.forEach((lottoTicket) => {
-        Console.print(`[${lottoTicket.join(`${DELIMETER} `)}]`);
-      });
-    }
-  }
-
-  async promptPurchaseAmount() {
-    try {
-      const newPurchaseAmount = await Console.readLineAsync(
-        InputPrompts.purchaseAmount,
-      );
-
-      this.setPurchaseAmount(newPurchaseAmount);
-    } catch (error) {
-      Console.print(`${error.message}\n`);
-      await this.promptPurchaseAmount();
-    }
-  }
-
-  async promptBonusNumber() {
-    try {
-      const bonusNumber = await Console.readLineAsync(InputPrompts.bonusNumber);
-
-      this.setBonusNumber(bonusNumber);
-    } catch (error) {
-      Console.print(`${error.message}\n`);
-      await this.promptBonusNumber();
     }
   }
 
@@ -124,42 +90,6 @@ class LottoController {
     });
 
     return winningResult;
-  }
-
-  PrintLottoWinningResult(winningResult) {
-    const winningResultMessage = [
-      '당첨 통계',
-      '---',
-      `3개 일치 (${Prize.MATCH_3}원) - ${winningResult[3]}개`,
-      `4개 일치 (${Prize.MATCH_4}원) - ${winningResult[4]}개`,
-      `5개 일치 (${Prize.MATCH_5}원) - ${winningResult[5]}개`,
-      `5개 일치, 보너스 볼 일치 (${Prize.MATCH_5_BONUS}원) - ${winningResult['5B']}개`,
-      `6개 일치 (${Prize.MATCH_6}원) - ${winningResult[6]}개`,
-    ];
-
-    winningResultMessage.forEach((message) => Console.print(message));
-  }
-
-  printRateOfReturn(winningResult) {
-    const totalPrizeMoney = Object.entries(winningResult).reduce(
-      (acc, [matchCount, lottoCount]) =>
-        acc + lottoCount * PrizeMoney[matchCount],
-      0,
-    );
-
-    const rateOfReturn = (
-      (totalPrizeMoney / this.#purchaseAmount) *
-      100
-    ).toFixed(1);
-
-    Console.print(OutputMessages.TOTAL_ROR(rateOfReturn));
-  }
-
-  displayWinningResult(winningNumbers) {
-    const winningResult = this.compareLottoTickets(winningNumbers);
-
-    this.PrintLottoWinningResult(winningResult);
-    this.printRateOfReturn(winningResult);
   }
 }
